Moved element-snapshot-formatter to common directory for all tests
